Remove "PRAGMA synchronous=OFF" to prevent "Safety level may not be changed inside a transaction" issue.
This issue was encountered when using SqliteSource.from_records(...)
to build a data source using records filtered from a generator
comprehension iterating over an existing CsvSource.  Warning: We were
unable to replicate the behavior in testing.  But setting
"synchronous=OFF" should have little-to-no effect on the speed of
SELECT queries so it's not important to include.