Change callable keyword handling so that user defined functions do not reuse names with the create_function() method. Instead, functions are now registered by id() value per connection.
The previous implementation registered new functions for every
query but minimized the number of total registered functions by
**reusing** the same function names ('<table>_FUNC0',
'<table>_FUNC1', etc.). But when creating several functions in
quick succession, the connection.create_function() was failing
with the message:

    Traceback (most recent call last):
      File "datatest/dataaccess.py", line 1081, in _execute_query
        self._connection.create_function(name, 1, func)
    sqlite3.OperationalError: Error creating function

This failure could not be replicated in a concise and fast test
case but was occuring consistently with production data sets. One
specific production case registered 30 functions on a data set
that was 23 columns x 280,000+ rows---not terribly large but too
large to quickly load/generate in testing.