Remove Selector.__del__() method to prevent race condition issues with very short-lived Selectors.
The __del__() method was running a "DROP TABLE ..."
command to remove unused disk space during the running
process. Unfortunately this was encountering database
locks and raising errors when Selectors were created and
immediately destroyed. Since the disk space is reclaimed
when the process finishes, it's not so important to
continually reclaim it while the process is still
running.

This error is possible because Selectors currently share
a single database connection that persists for the entire
process. There are a few ways to address this in the future:

1. Use individual database connections for each Selector
   to allow error free disk-space reclamation during
   the process at the cost of 1 file descriptor per
   Selector instance (in the case of temp-file connections).

2. Add a close() or drop() method so that users can choose
   to reclaim space if it becomes a problem but allow usage
   to accumulate when it's not a problem.

3. Do nothing--allow disk usage to grow as Selectors are
   created and let them get cleaned up when the process
   finishes.

Option 1 seems to be the most appealing and if there is an
issue with excess file descriptors, we could add a flag to
trigger ":memory:" connections or users could define a
MemorySelector or SharedConnectionSelector.