Add a check for missing or older versions of SQLite 3.
Datatest's Selector class uses the "sqlite3" package
from the Python Standard Library. But some builds of
Python don't include the entire library and others
were bundled with out-of-date versions of SQLite.

The Selector requires SQLite 3.6.8 (released in January
of 2009). If SQLite is missing or too old, an error is
raised explaining this. Some cases where this is an
issue:

* Windows binaries for Python 3.1.4 and 2.6.6 were
  bundled with SQLite 3.5.9.
* Jython 2.7 and 2.5 don't bundle "sqlite3" in their
  distributions at all--batteries not included!