Rename Query method from ``__call__`` to ``execute``.
Since the Query and Selector classes are closely coupled,
new users might lose track of object type they're working
with. And when both of these objects are callable but use
different signatures the failure message could be very
confusing. By renaming the Query's method, we help to
differentiate these two objects and clarify what it meant
to call a Query object.