Renamed the bundled pytest plugin from pytest_datatest to _pytest_datatest.
With the release of Pytest 3.6.0, pluggy--a pytest dependency
that was recently updated--has been raising the following
error:

  pluggy.PluginValidationError: unknown hook 'pytest_datatest'
  in plugin <module 'datatest._pytest_plugin'

The related behavior was changed in Pytest 3.6.0 in response
to a deprecation in pluggy (pytest-dev/pytest#3487). While
it seems clear that pluggy supported this without problems
(see pytest-dev/pluggy#91), I'm not sure if the latest Pytest
behavior is intentional or a regression. But in either case,
we want to make sure the above error is avoided.