Register tuple as a BaseElement (add to subclass-hook method).
This formalizes the treatment of tuples as single records
rather than as collections of items that should be processed
individually. Previously, tuple handling was less consistent
and couldn't be easily fixed because many functions were
implementing automatic argument unpacking which, while
convenient for simple cases, led to confusing complexity
in more complex queries.