Add __slots__ to reduce memory consumption and prevent dynamic attribute creation.
Adding __slots__ to save memory is an idea I've entertained for a while
as there are cases where a lot of differences could be created at the
same time. But limiting attribute creation has the side benefit of helping
instances act more like immutable objects--necessary for hashable objects.
Now that I have two reasons to add __slots__, I guess it's time to
implement them.

Note: Preventing dynamic attribute creation with __setattr__() is not
an attractive option because it makes the class more complex and doesn't
come with the benefit of memory savings.