Change module name and language from "allowance" to "acceptance".
Some background:

Initially there was very little overlap between the behavior of validation
methods and "allowance" methods. But as the methods have been refined and
expanded, we now have cases like validate.approx/allowed.deviation and
validate.fuzzy/allowed.fuzzy. The validation method should be used when
the variation is intentional and the "allowance" should be used when the
variation is unintentional.

This use of "allowance" for unintentional variation is counter to
existing manufacturing and engineering vocabulary. But the word choice
was made intentionally because "accept" is sounds very similar to the
Python reserved word "except" and I wanted to clearly separate these two
concepts to facilitate discussing test code.

The growing overlap between validation methods and "allowance" methods
has brought this issue to the foreground. The need to clarify their use
is now important enough that I think it's worth having the near-homonym
of "accept" and "except".