Invert the semantics of subset() and superset() methods.
This behavior was flipped so that it more closely match how
users assume these methods will behave. The previous semantics
were used because they more accurately reflect the internal
structure of datatest. But those are implementation details
that aren't as important as having the API more closely match
user expectations.