Deprecate the skip(), skipIf(), and skipUnless() decorators.
These decorators were originally created back when Python 3.2 was
in common use. In 3.2, the equivalent decorators from unittest did
not return an object with a `__wrapped__` attribute. In order to
maintain incremental testing order (which requires introspecting
the test methods), datatest implemented its own "skip" decorators.

Since Python 3.3, functools.wraps() includes a __wrapped__
attribute that solves the original problem.

Python 3.2 reached end-of-life in February of 2016. Although
datatest still supports 3.2, the sort behavior of skipped tests
is not critical for end-of-life versions of Python.